test: refactor use __test__ folder, use describe blocks in versioncheck
