+++ Source/Plugins/Core/com.equella.core/js/__tests__/versioncheck.test.js
@@ -8,36 +8,40 @@ const version_check_js_file = fs.readFileSync(
 );
 eval(version_check_js_file);
 
-test("Test major updates", () => {
-  const checkResult = createCheckResult("2018.1.0", mockData.mockReleases);
-  expect(checkResult.newer).toBe(true);
-  expect(checkResult.newerReleases.majorUpdate.major).toBe("2019");
-});
+describe("versioncheck", () => {
+  describe("createCheckResult", () => {
+    it("should support major updates", () => {
+      const checkResult = createCheckResult("2018.1.0", mockData.mockReleases);
+      expect(checkResult.newer).toBe(true);
+      expect(checkResult.newerReleases.majorUpdate.major).toBe("2019");
+    });
 
-test("Test minor updates", () => {
-  const checkResult = createCheckResult("2019.1.1", mockData.mockReleases);
-  expect(checkResult.newer).toBe(true);
+    it("should support minor updates", () => {
+      const checkResult = createCheckResult("2019.1.1", mockData.mockReleases);
+      expect(checkResult.newer).toBe(true);
 
-  const minorUpdateRelease = checkResult.newerReleases.minorUpdate;
-  expect(minorUpdateRelease.major).toBe("2019");
-  expect(minorUpdateRelease.minor).toBe("2");
-});
+      const minorUpdateRelease = checkResult.newerReleases.minorUpdate;
+      expect(minorUpdateRelease.major).toBe("2019");
+      expect(minorUpdateRelease.minor).toBe("2");
+    });
 
-test("Test patch updates", () => {
-  const checkResult = createCheckResult("2019.2.0", mockData.mockReleases);
-  expect(checkResult.newer).toBe(true);
+    it("should support patch updates", () => {
+      const checkResult = createCheckResult("2019.2.0", mockData.mockReleases);
+      expect(checkResult.newer).toBe(true);
 
-  const patchUpdateRelease = checkResult.newerReleases.patchUpdate;
-  expect(patchUpdateRelease.major).toBe("2019");
-  expect(patchUpdateRelease.minor).toBe("2");
-  expect(patchUpdateRelease.patch).toBe("1");
-});
+      const patchUpdateRelease = checkResult.newerReleases.patchUpdate;
+      expect(patchUpdateRelease.major).toBe("2019");
+      expect(patchUpdateRelease.minor).toBe("2");
+      expect(patchUpdateRelease.patch).toBe("1");
+    });
 
-test("Test no updates", () => {
-  const checkResult = createCheckResult("2020.1.0", mockData.mockReleases);
-  expect(checkResult.newer).toBe(false);
+    it("should do nothing when there are no updates", () => {
+      const checkResult = createCheckResult("2020.1.0", mockData.mockReleases);
+      expect(checkResult.newer).toBe(false);
 
-  expect(checkResult.newerReleases.majorUpdate).toBe(null);
-  expect(checkResult.newerReleases.minorUpdate).toBe(null);
-  expect(checkResult.newerReleases.patchUpdate).toBe(null);
+      expect(checkResult.newerReleases.majorUpdate).toBe(null);
+      expect(checkResult.newerReleases.minorUpdate).toBe(null);
+      expect(checkResult.newerReleases.patchUpdate).toBe(null);
+    });
+  });
 });
