+++ Source/Plugins/Core/com.equella.core/js/tsrc/__tests__/util/commonstrings.test.ts
@@ -1,4 +1,4 @@
-import { commonString } from "./commonstrings";
+import { commonString } from "../../util/commonstrings";
 
 describe("commonstrings", () => {
   describe("commonString", () => {
