+++ Source/Plugins/Core/com.equella.core/js/tsrc/__tests__/util/dictionary.test.ts
@@ -1,4 +1,4 @@
-import { properties } from "./dictionary";
+import { properties } from "../../util/dictionary";
 
 describe("dictionary", () => {
   describe("properties", () => {
