+++ Source/Plugins/Core/com.equella.core/js/tsrc/__tests__/util/encodequery.test.ts
@@ -1,4 +1,4 @@
-import { encodeQuery } from "./encodequery";
+import { encodeQuery } from "../../util/encodequery";
 
 describe("encodequery", () => {
   describe("encodeQuery", () => {
